+++ nflow-engine/src/main/java/io/nflow/engine/config/PropertiesConfiguration.java
@@ -1,4 +1,4 @@
-package io.nflow.engine.internal.config;
+package io.nflow.engine.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
