+++ nflow-engine/src/main/java/io/nflow/engine/internal/storage/db/OracleDatabaseConfiguration.java
@@ -1,6 +1,6 @@
 package io.nflow.engine.internal.storage.db;
 
-import static io.nflow.engine.internal.config.Profiles.ORACLE;
+import static io.nflow.engine.config.Profiles.ORACLE;
 import static org.slf4j.LoggerFactory.getLogger;
 
 import java.sql.Connection;
@@ -19,7 +19,7 @@ import org.springframework.core.env.Environment;
 import org.springframework.jdbc.datasource.DataSourceUtils;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import io.nflow.engine.internal.config.NFlow;
+import io.nflow.engine.config.NFlow;
 import io.nflow.engine.workflow.instance.WorkflowInstance.WorkflowInstanceStatus;
 
 @Profile(ORACLE)
