+++ nflow-engine/src/test/java/io/nflow/engine/config/PropertiesConfigurationTest.java
@@ -1,8 +1,10 @@
-package io.nflow.engine.internal.config;
+package io.nflow.engine.config;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.junit.MockitoJUnitRunner;
+
+import io.nflow.engine.config.PropertiesConfiguration;
 
 @RunWith(MockitoJUnitRunner.class)
 public class PropertiesConfigurationTest {
